update checkout and status to reflect
Instead of individual files, checkout now moves entire directories to the trash if all sub-items of the directory are deleted. This ensures that package bundles like *.pxd are kept as a bundle in the trash. It also makes restoring files and directories easier. getStatus was updated as well to reflect the changes needed by checkout